+++ pangea-sdk/internal/pangeatesting/integration.go
@@ -6,57 +6,45 @@ import (
 	"testing"
 	"time"
 
-	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/internal/defaults"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/internal/pangeautil"
+	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/option"
 	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/pangea"
 )
 
 func IntegrationConfig(t *testing.T, env TestEnvironment) *pangea.Config {
-	return &pangea.Config{
-		HTTPClient:         defaults.HTTPClient(),
-		Domain:             GetTestDomain(t, env),
-		Token:              GetTestToken(t, env),
-		QueuedRetryEnabled: true,
-		PollResultTimeout:  60 * time.Second,
-		Retry:              true,
-		Logger:             pangeautil.GetDebugLogger(),
-		RetryConfig: &pangea.RetryConfig{
-			RetryMax: 4,
-		},
+	config, err := pangea.NewConfig(
+		option.WithDomain(GetTestDomain(t, env)),
+		option.WithLogger(pangeautil.GetDebugLogger()),
+		option.WithPollResultTimeout(60*time.Second),
+		option.WithQueuedRetryEnabled(true),
+		option.WithToken(GetTestToken(t, env)),
+	)
+	if err != nil {
+		panic(err)
 	}
+	return config
 }
 
 func IntegrationAuditVaultConfig(t *testing.T, env TestEnvironment) *pangea.Config {
-	return &pangea.Config{
-		HTTPClient:         defaults.HTTPClient(),
-		Domain:             GetTestDomain(t, env),
-		Token:              GetVaultSignatureTestToken(t, env),
-		QueuedRetryEnabled: true,
-		PollResultTimeout:  60 * time.Second,
-		Retry:              true,
-		Logger:             pangeautil.GetDebugLogger(),
-		RetryConfig: &pangea.RetryConfig{
-			RetryMax: 4,
-		},
+	config := IntegrationConfig(t, env)
+	if err := config.Apply(option.WithToken(GetVaultSignatureTestToken(t, env))); err != nil {
+		panic(err)
 	}
+	return config
 }
 
 func IntegrationCustomSchemaConfig(t *testing.T, env TestEnvironment) *pangea.Config {
-	return &pangea.Config{
-		HTTPClient:         defaults.HTTPClient(),
-		Domain:             GetTestDomain(t, env),
-		Token:              GetCustomSchemaTestToken(t, env),
-		PollResultTimeout:  60 * time.Second,
-		QueuedRetryEnabled: true,
-		Logger:             pangeautil.GetDebugLogger(),
+	config := IntegrationConfig(t, env)
+	if err := config.Apply(option.WithToken(GetCustomSchemaTestToken(t, env))); err != nil {
+		panic(err)
 	}
+	return config
 }
 
 func IntegrationMultiConfigConfig(t *testing.T, env TestEnvironment) *pangea.Config {
-	return &pangea.Config{
-		HTTPClient: defaults.HTTPClient(),
-		Domain:     GetTestDomain(t, env),
-		Token:      GetMultiConfigTestToken(t, env),
-		Logger:     pangeautil.GetDebugLogger(),
+	config := IntegrationConfig(t, env)
+	if err := config.Apply(option.WithToken(GetMultiConfigTestToken(t, env))); err != nil {
+		panic(err)
 	}
+	return config
 }